// packages/engine/src/avatar/types.ts
export enum HumanBone {
  Chest = "chest",
  Head = "head",
  Hips = "hips",
  Jaw = "jaw",
  LeftEye = "leftEye",
  LeftFoot = "leftFoot",
  LeftHand = "leftHand",
  LeftIndexDistal = "leftIndexDistal",
  LeftIndexIntermediate = "leftIndexIntermediate",
  LeftIndexProximal = "leftIndexProximal",
  LeftLittleDistal = "leftLittleDistal",
  LeftLittleIntermediate = "leftLittleIntermediate",
  LeftLittleProximal = "leftLittleProximal",
  LeftLowerArm = "leftLowerArm",
  LeftLowerLeg = "leftLowerLeg",
  LeftMiddleDistal = "leftMiddleDistal",
  LeftMiddleIntermediate = "leftMiddleIntermediate",
  LeftMiddleProximal = "leftMiddleProximal",
  LeftRingDistal = "leftRingDistal",
  LeftRingIntermediate = "leftRingIntermediate",
  LeftRingProximal = "leftRingProximal",
  LeftShoulder = "leftShoulder",
  LeftThumbDistal = "leftThumbDistal",
  LeftThumbIntermediate = "leftThumbIntermediate",
  LeftThumbProximal = "leftThumbProximal",
  LeftToes = "leftToes",
  LeftUpperArm = "leftUpperArm",
  LeftUpperLeg = "leftUpperLeg",
  Neck = "neck",
  RightEye = "rightEye",
  RightFoot = "rightFoot",
  RightHand = "rightHand",
  RightIndexDistal = "rightIndexDistal",
  RightIndexIntermediate = "rightIndexIntermediate",
  RightIndexProximal = "rightIndexProximal",
  RightLittleDistal = "rightLittleDistal",
  RightLittleIntermediate = "rightLittleIntermediate",
  RightLittleProximal = "rightLittleProximal",
  RightLowerArm = "rightLowerArm",
  RightLowerLeg = "rightLowerLeg",
  RightMiddleDistal = "rightMiddleDistal",
  RightMiddleIntermediate = "rightMiddleIntermediate",
  RightMiddleProximal = "rightMiddleProximal",
  RightRingDistal = "rightRingDistal",
  RightRingIntermediate = "rightRingIntermediate",
  RightRingProximal = "rightRingProximal",
  RightShoulder = "rightShoulder",
  RightThumbDistal = "rightThumbDistal",
  RightThumbIntermediate = "rightThumbIntermediate",
  RightThumbProximal = "rightThumbProximal",
  RightToes = "rightToes",
  RightUpperArm = "rightUpperArm",
  RightUpperLeg = "rightUpperLeg",
  Spine = "spine",
  UpperChest = "upperChest",
}